refactor: move pages to own folders, move createManagedWidget
