+++ jupyterlab_caip_optimizer/src/components/create_study/index.spec.tsx
@@ -1,16 +1,16 @@
 import * as React from 'react';
-import { createDropdown, DropdownItem, CreateStudy } from './create_study';
+import { createDropdown, DropdownItem, CreateStudy } from '.';
 import { rest } from 'msw';
 import { setupServer } from 'msw/node';
-import * as Types from '../types';
+import * as Types from '../../types';
 import {
   screen,
   render,
   waitFor,
   waitForElementToBeRemoved,
-} from '../utils/redux_render';
-import { proxyUrl, createStudyUrl } from '../utils/urls';
-import { fakeStudy, cleanFakeStudyName } from '../service/test-constants';
+} from '../../utils/redux_render';
+import { proxyUrl, createStudyUrl } from '../../utils/urls';
+import { fakeStudy, cleanFakeStudyName } from '../../service/test-constants';
 import userEvent from '@testing-library/user-event';
 
 const server = setupServer();
