+++ jupyterlab_caip_optimizer/src/components/sidebar_widget/index.tsx
@@ -28,14 +28,14 @@ import {
   TableRow,
   Typography,
 } from '@material-ui/core';
-import { RootState, AppDispatch } from '../store/store';
-import { Study } from '../types';
-import { setView } from '../store/view';
+import { RootState, AppDispatch } from '../../store/store';
+import { Study } from '../../types';
+import { setView } from '../../store/view';
 import { Store } from 'redux';
-import { prettifyStudyName } from '../service/optimizer';
+import { prettifyStudyName } from '../../service/optimizer';
 import { style } from 'typestyle';
-import { styles } from '../utils/styles';
-import { Loading } from './loading';
+import { styles } from '../../utils/styles';
+import { Loading } from '../misc/loading';
 
 const rowStyle = style({
   $nest: {
