+++ jupyterlab_caip_optimizer/src/index.ts
@@ -4,10 +4,7 @@ import {
   JupyterFrontEnd,
   JupyterFrontEndPlugin,
 } from '@jupyterlab/application';
-import { Widget } from '@phosphor/widgets';
-import { Store } from 'redux';
 import { store } from './store/store';
-import { watch } from './store/watch';
 import { MainAreaWidget } from './components/main_area_widget';
 import { SidebarWidget } from './components/sidebar_widget';
 import { fetchStudies } from './store/studies';
@@ -15,51 +12,9 @@ import { fetchMetadata } from './store/metadata';
 import { createSnack } from './store/snackbar';
 import { SnackbarWidget } from './components/snackbar_widget';
 import { unwrapResult } from '@reduxjs/toolkit';
-
-/**
- * Opens and closes a widget based on redux store's `view.isVisible` property.
- * @param reduxStore the redux store with `view.isVisible`.
- * @param app the jupyter application.
- * @param widgetComponent the widget to be managed by redux.
- */
-const createManagedWidget = <
-  WIDGET extends Widget,
-  STORE extends Store<{ view: { isVisible: boolean } }>
->(
-  reduxStore: STORE,
-  app: JupyterFrontEnd,
-  widgetComponent: new (reduxStore: STORE) => WIDGET
-) => {
-  let widget = new widgetComponent(reduxStore);
-  widget.id = 'optimizer:main-area';
-
-  const onChange = watch(
-    reduxStore.getState,
-    state => state.view.isVisible,
-    (previousIsVisible, nextIsVisible) => previousIsVisible === nextIsVisible
-  );
-  if (reduxStore.getState().view.isVisible) {
-    app.shell.add(widget, 'main');
-    app.shell.activateById(widget.id);
-  }
-  reduxStore.subscribe(
-    onChange(isVisible => {
-      if (isVisible) {
-        app.shell.add(widget, 'main');
-        app.shell.activateById(widget.id);
-      } else {
-        widget.close();
-        // Setup component for next open
-        widget = new widgetComponent(reduxStore);
-        widget.id = 'optimizer:main-area';
-      }
-    })
-  );
-};
+import { createManagedWidget } from './utils/create_managed_widget';
 
 async function activate(app: JupyterFrontEnd) {
-  store.dispatch(fetchStudies());
-
   // Create main area widget
   createManagedWidget(store, app, MainAreaWidget);
 
@@ -68,6 +23,7 @@ async function activate(app: JupyterFrontEnd) {
   app.shell.add(sidebarWidget, 'left', { rank: 100 });
   app.shell.add(new SnackbarWidget(store), 'header');
 
+  // Fetch Metadata for project id, region and then load studies
   await store
     .dispatch(fetchMetadata())
     .then(() => store.dispatch(fetchStudies()))
